requirements.txt
selenium
webdriver-manager
python-telegram-bot==20.3